make play deterministic
